+++ microhapulator/tests/test_pipe.py
@@ -49,8 +49,9 @@ def test_pipe_gbr_usc10(tmp_path):
         "gbr-usc",
         "--workdir",
         str(tmp_path),
-        "--threads",
-        "1",
+        "--threads=1",
+        "--static=5",
+        "--dynamic=0.02",
         "--hg38",
         hg38,
     ]
@@ -63,4 +64,9 @@ def test_pipe_gbr_usc10(tmp_path):
     assert (tmp_path / "report.html").is_file()
     expected = pd.read_csv(data_file("gbr-usc-summary.tsv"), sep="\t")
     observed = pd.read_csv(tmp_path / "analysis" / "summary.tsv", sep="\t")
+    assert observed.equals(expected)
+    profile = tmp_path / "analysis" / "gbr-usc" / "profiles" / "gbr-usc-quant.csv"
+    assert profile.is_file()
+    expected = pd.read_csv(data_file("gbr-usc-profile.csv"))
+    observed = pd.read_csv(profile)
     assert observed.equals(expected)