+++ CHANGELOG.md
@@ -6,10 +6,11 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 - Profiles compatible with probgen programs now included in pipe output (#135).
-- Haplotype call plots now included in the HTML report (#136).
+- Haplotype call plots now included in the pipe HTML report (#136).
 
 ### Changed
 - Exposed static and dynamic threshold configuration to pipe CLI (#135).
+- Updated plot colors in the pipe HTML report (#139).
 
 
 ## [0.6.1] 2022-04-27