+++ CHANGELOG.md
@@ -7,13 +7,13 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - Profiles compatible with probgen programs now included in pipe output (#135).
 - Haplotype call plots now included in the pipe HTML report (#136).
-- Added offtarget module to count reads that map to off target loci in hg38 (#143).
+- Added offtarget module to count reads that map to off target loci in hg38 (#143, #153).
 - Added typing rate and mapping rate information per marker to the main pipe HTML report (#146).
 - Added marker detail HTML report (#146, #151).
 - Implemented support for single-end reads in the end-to-end microhap analysis pipeline (#147).
 
 ### Changed
-- Exposed static and dynamic threshold configuration to pipe CLI (#135).
+- Exposed static and dynamic threshold configuration to pipe CLI (#135, #153).
 - Updated plot colors in the pipe HTML report (#139).
 - Updated pipe HTML report to conditionally plot read length histograms or tables depending on uniformity (#140).